% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addLegend}
\alias{addLegend}
\title{Add Legend}
\usage{
addLegend(legend.loc = "center", ncol = 1, ...)
}
\arguments{
\item{legend.loc}{legend.loc places a legend into one of nine locations on
the chart: bottomright, bottom, bottomleft, left, topleft, top, topright,
right, or center.}

\item{ncol}{number of columns for the legend}

\item{\dots}{any other passthrough parameters. Not currently used.}
}
\description{
Add Legend
}
\author{
Ross Bennett
}

