\name{indexClass}
\alias{indexClass}
\title{ Extracting the Class of an xtsdf Index }
\description{
Generic functions to extract, replace, and format the class of the index
of an xtsdf object.
}
\usage{
indexClass(x)
}
\arguments{
  \item{x}{ xts or xtsdf object }
}
\details{Provides a generic over the \code{indexClass} function of \code{xts}.}
\author{ Michael Weylandt }
